feat(telemetry): Loki and Otel loggers
